tugetha-payments/requirements.txt
django==4.2
djangorestframework==3.15
djangorestframework-simplejwt==5.3
django-environ==0.11
django-cors-headers==4.3
psycopg2-binary==2.9
celery==5.3
redis==5.0
firebase-admin==6.4
requests==2.31
gunicorn==21.2
whitenoise==6.6
django-extensions==3.2
